pyproject.toml
[project]
name = "data-gent"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "atlassian-python-api>=3.41.19",
    "db-dtypes>=1.4.2",
    "dotenv>=0.9.9",
    "fastapi>=0.115.11",
    "google-cloud-bigquery>=3.30.0",
    "google-cloud-storage>=3.1.0",
    "ipython>=9.0.2",
    "langgraph>=0.3.9",
    "langgraph-cli[inmem]>=0.1.77",
    "openai>=1.65.5",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "python-mermaid>=0.1.3",
    "toml>=0.10.2",
    "uvicorn>=0.34.0",
]
